+++ llm_ffw/__init__.py
@@ -8,6 +8,7 @@ from .capabilities import (
     IPAddressCapability,
     MACAddressCapability,
     IBANCapability,
+    AuthorizationHeaderCapability,
     RuleCapability,
     SecretCatalogCapability,
     UnsafeURLCapability,
@@ -39,6 +40,7 @@ from .json_output import JSONOutputConfig
 from .ip_address import IPAddressConfig
 from .mac_address import MACAddressConfig
 from .iban import IBANConfig
+from .authorization_header import AuthorizationHeaderConfig
 from .email_address import EmailAddressConfig
 from .unsafe_url import UnsafeURLConfig
 from .payment_card import PaymentCardConfig
@@ -89,6 +91,7 @@ from .rules.json_output import JSONOutputRule
 from .rules.ip_address import IPAddressRule
 from .rules.mac_address import MACAddressRule
 from .rules.iban import IBANRule
+from .rules.authorization_header import AuthorizationHeaderRule
 from .rules.email_address import EmailAddressRule
 from .rules.unsafe_url import UnsafeURLRule
 from .rules.payment_card import PaymentCardRule
@@ -135,6 +138,9 @@ __all__ = [
     "IBANCapability",
     "IBANConfig",
     "IBANRule",
+    "AuthorizationHeaderCapability",
+    "AuthorizationHeaderConfig",
+    "AuthorizationHeaderRule",
     "MACAddressConfig",
     "MACAddressRule",
     "LLMFirewall",